// src/app/v2/citizen/dto/citizen.dto.ts
export class CitizenDto {
    fullName?: string;
    cpf?: string;
    cns?: string;
    birthDate?: Date;
    gender?: string;
    age?: number;
    weighting?: number;
    identificationType?: string;
    lastContactDate?: Date;
    totalServices?: number;
    city?: string;
    district?: string;
}